chore(events): change package structure (move users code to event package) #37
