+++ backend/src/test/java/unischedule/users/service/UsersServiceImplTest.java
@@ -17,6 +17,7 @@ import unischedule.events.dto.EventCreateResponseDto;
 import unischedule.events.dto.EventGetResponseDto;
 import unischedule.events.entity.Event;
 import unischedule.events.repository.EventRepository;
+import unischedule.events.service.EventService;
 
 @ExtendWith(MockitoExtension.class) // Mockito 초기화
 class UsersServiceImplTest {
@@ -24,7 +25,7 @@ class UsersServiceImplTest {
     private EventRepository eventRepository;
     
     @InjectMocks
-    private UsersService usersService;
+    private EventService eventService;
     
     @Test
     void testMakeEvent() {
@@ -37,7 +38,7 @@ class UsersServiceImplTest {
         );
         
         Event savedEvent = new Event(
-            1L, "회의", "주간 회의",
+            "회의", "주간 회의",
             requestDto.startTime(), requestDto.endTime(),
             "CONFIRMED", true
         );
@@ -45,12 +46,12 @@ class UsersServiceImplTest {
         Mockito.when(eventRepository.save(any(Event.class)))
             .thenReturn(savedEvent);
         
-        Mockito.when(eventRepository.existsByCreatorIdAndStartAtLessThanAndEndAtGreaterThan(eq(1L), any(LocalDateTime.class), any(
+        Mockito.when(eventRepository.existsByStartAtLessThanAndEndAtGreaterThan(any(LocalDateTime.class), any(
                 LocalDateTime.class)))
             .thenReturn(false);
         
         // when
-        EventCreateResponseDto result = usersService.makeEvent(1L, requestDto);
+        EventCreateResponseDto result = eventService.makeEvent(requestDto);
         
         // then
         assertThat(result.eventId()).isEqualTo(1L);
@@ -65,25 +66,24 @@ class UsersServiceImplTest {
         Long userId = 1L;
         
         Event event1 = new Event(
-            1L, "회의", "주간 회의",
+            "회의", "주간 회의",
             LocalDateTime.of(2025, 9, 10, 10, 0),
             LocalDateTime.of(2025, 9, 10, 11, 0),
             "CONFIRMED", true
         );
         
-        Event event2 = new Event(
-            1L, "워크샵", "분기별 워크샵",
+        Event event2 = new Event("워크샵", "분기별 워크샵",
             LocalDateTime.of(2025, 9, 15, 14, 0),
             LocalDateTime.of(2025, 9, 15, 17, 0),
             "CONFIRMED", false
         );
         
         // Mockito로 findByStartAtGreaterThanEqualAndEndAtLessThanEqual 호출 시 반환값 지정
-        Mockito.when(eventRepository.findByCreatorIdAndStartAtLessThanAndEndAtGreaterThan(userId, end, start))
+        Mockito.when(eventRepository.findByStartAtLessThanAndEndAtGreaterThan(end, start))
             .thenReturn(List.of(event1, event2));
         
         // when
-        List<EventGetResponseDto> result = usersService.getEvents(start, end, userId);
+        List<EventGetResponseDto> result = eventService.getEvents(start, end);
         
         // then
         assertThat(result).hasSize(2);
